removed examples directory as there are no examples...
