+++ rspoectl.cpp
